Revert "github: Add release workflow"
This reverts commit 9a360bcb8e892188880e49aa10caca23928c556e.

After some digging in the source code of the action I've realized
that since this will only update major tags, an update from `1.2.3` to
`1.2.4` will result in a stale `1.2` tag, but an up to date `1` tag.

Since this requires manual intervention anyway, we may as well do both
bumps manually for now. Perhaps if we switch to branches for the major
and minor versions this won't an issue. This will require some testing.